asm: rename the AMD64 into X86_64
in this context, AMD64 really means amd64 or em64t, so let's
rename this into X86_64 in order to avoid any confusion


(back-ported from commit open-mpi/ompi@af0b5cffb4a013c97089bdc7e26d6a058357f8c1)
